+++ ui/app/routes/jobs/job/versions.js
@@ -1,6 +1,6 @@
 import Route from '@ember/routing/route';
 import { collect } from '@ember/object/computed';
-import { watchRelationship } from 'nomad-ui/utils/properties/watch';
+import { watchRecord, watchRelationship } from 'nomad-ui/utils/properties/watch';
 import WithWatchers from 'nomad-ui/mixins/with-watchers';
 
 export default class VersionsRoute extends Route.extend(WithWatchers) {
@@ -11,10 +11,13 @@ export default class VersionsRoute extends Route.extend(WithWatchers) {
 
   startWatchers(controller, model) {
     if (model) {
-      controller.set('watcher', this.watchVersions.perform(model));
+      controller.set('watcher', this.watch.perform(model));
+      controller.set('watchVersions', this.watchVersions.perform(model));
     }
   }
 
+  @watchRecord('job') watch;
   @watchRelationship('versions') watchVersions;
-  @collect('watchVersions') watchers;
+
+  @collect('watch', 'watchVersions') watchers;
 }