+++ ui/tests/pages/layout.js
@@ -99,4 +99,14 @@ export default create({
     title: text('[data-test-error-title]'),
     message: text('[data-test-error-message]'),
   },
+
+  inlineError: {
+    isShown: isPresent('[data-test-inline-error]'),
+    title: text('[data-test-inline-error-title]'),
+    message: text('[data-test-inline-error-body]'),
+    dismiss: clickable('[data-test-inline-error-close]'),
+
+    isDanger: hasClass('is-danger', '[data-test-inline-error]'),
+    isWarning: hasClass('is-warning', '[data-test-inline-error]'),
+  },
 });